fix: version deployments modal refactor
